WelfordAccumulator to online_moments.hpp
